cleanup deprecated subcommands
